handler use a simple file
